requirements.txt
tritonclient[grpc]
orjson==3.10.2
open-clip-torch==2.29.0
hf_transfer==0.1.8
torch==2.6.0 --extra-index-url https://download.pytorch.org/whl/cpu
torchvision==0.21.0 --extra-index-url https://download.pytorch.org/whl/cpu
pillow==11.1.0
pycurl==7.45.3
fastapi==0.115.2
uvicorn==0.34.0
validators==0.34.0
transformers==4.48.2
sentence_transformers==3.4.1
tokenizers==0.21.0
pydantic==2.8.2
certifi
onnx==1.17.0
timm==1.0.14